backend/requirements.txt
flask==0.12.2
flask-cors==3.0.3
Pillow==4.3.0
numpy==1.12.0
scipy==1.0.0